pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "coalmine"
version = "0.1.0"
description = "Cloud Canary Token Management - Deploy, monitor, and rotate deceptive credentials across AWS and GCP"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "John Earle"}
]
keywords = ["security", "canary", "honeypot", "aws", "gcp", "cloud"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Security",
]
dependencies = [
    "celery[redis]",
    "sqlalchemy",
    "psycopg2-binary",
    "alembic",
    "python-dotenv",
    "boto3",
    "google-cloud-logging",
    "google-cloud-storage",
    "google-cloud-resource-manager",
    "pyyaml",
    "pydantic>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]",
    "python-multipart",  # Required for form data in session auth
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]

[project.scripts]
coalmine = "src.cli:run"

[project.urls]
Homepage = "https://github.com/yourorg/coalmine"
Repository = "https://github.com/yourorg/coalmine"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
